feat: PostPage와 CommunityPage url경로 수정
